+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## Unreleased
 
-* If HAS_GPS==0 do now try to connect to GPS (#11)
+* If HAS_GPS==0 do not try to connect to GPS (#11)
 * Default value for HAS_GPS is now based on GPS_DEV
+* Image for armv6l architecture
 
 ## 2.1.0 (2023-08-20)